+++ doc/Changelog.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## V1.0.7 (2023-05-12)
+
+#### Features
+
+- 支持 win64 系统自动安装 chromium 浏览器
+- device.launch 支持 `args`、`ignoreDefaultArgs`参数
+
 ## V1.0.5 (2023-04-16)
 
 ### Bug Fixes